feat: add description textarea
