+++ package.json
@@ -15,10 +15,12 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.2.0",
     "@testing-library/user-event": "^13.5.0",
+    "@types/autosize": "^4.0.1",
     "@types/jest": "^27.5.0",
     "@types/node": "^16.11.33",
     "@types/react": "^18.0.8",
     "@types/react-dom": "^18.0.3",
+    "autosize": "^5.0.1",
     "axios": "^0.27.2",
     "classnames": "^2.3.1",
     "i18next": "^21.8.2",
