+++ src/views/Board/Board.tsx
@@ -26,7 +26,7 @@ export const Board: FC = () => {
       {boardData && (
         <div className="flex flex-col grow gap-2 h-full overflow-hidden">
           <div className="rounded-sm text-lg">
-            <span>{boardData.title}</span>
+            <span className="font-semibold">{boardData.title}</span>
           </div>
           <Columns columns={boardData.columns} boardId={params['boardId'] || ''} />
         </div>
